requirements.txt
bottle==0.12.25
proxy-tools==0.1.0
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Security==10.0
pyobjc-framework-WebKit==10.0
pywebview==4.4.1
SQLAlchemy==2.0.23
typing_extensions==4.8.0
